requirements.txt
Django==1.8.2
factory_boy==2.5.2